Added path model and updated synch hashes
Hash object no longer contains document's url directly.
Path is introduced as a new model which contains document's filesystem path and url. We assumed that filesystem path and edition are unique together.
Updated sync hahses to save and load path instances. When creating hashes, references to path need to be set. The current solution is not the most elegant one, but works. Some refactoring is still required.