Remove chrome driver which is no longer being used
We are no longer using the chrome driver since it's sometimes
crashing while handling big repositroies. It makes no sense
to have to configure in order to test locally/on CI while
it is not even being used. We can add it again if needed.

Also update change attribute name of the git repository class
and add pytest-django requirement