Update python (#55)
* refact: bump pytest

*As part of py310 migration, update pytest version. Include py310 classifier

* feat: update classifiers

*Drop py35, support py>=36

* feat: add .github workflow

* refact: conditionally install taf based on py version

* Remove chrome driver which is no longer being used

We are no longer using the chrome driver since it's sometimes
crashing while handling big repositroies. It makes no sense
to have to configure in order to test locally/on CI while
it is not even being used. We can add it again if needed.

Also update change attribute name of the git repository class
and add pytest-django requirement

* CI fixes

- Due to a TAF issue, tests don't work when TAF wheel is installed,
so clone and install it for now
- config git info (just to avoid git errors)

Co-authored-by: Renata <rvaderna@openlawlib.org>